Address compiler warning by comparing two different enum values as ints
git-svn-id: svn://svn.code.sf.net/p/marathon/code/trunk@1820 a75d40c9-0969-4759-85e5-97cb871b7ed4

+++ Source_Files/Sound/sound_macintosh.h
@@ -275,7 +275,7 @@ static void initialize_machine_sound_manager(
 {
 	OSErr error;
 
-	assert(kFullVolume==MAXIMUM_SOUND_VOLUME);
+	assert(static_cast<uint32>(kFullVolume)==static_cast<uint32>(MAXIMUM_SOUND_VOLUME));
 
  	// Thomas Herzog: Updated for Universal Interfaces 3.4
  	#if UNIVERSAL_INTERFACES_VERSION < 0x0340
